Finish
Organice en código según su tipo
PerfilScreen : Solo contiene estructura y lógica.
PerfilHeader : Avatar + Nombre + Botón editar.
OptionCard :Cada tarjeta de opción.
EditProfileBottomSheet: Formulario para editar perfil.
