+++ frontend/zona44app/lib/pages/Home/home.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:zona44app/exports/exports.dart';
 import 'package:zona44app/pages/Home/bloc/home_bloc.dart';
+import 'package:zona44app/screens/perfil_screen.dart';
 
 // Página principal que maneja la navegación entre Inicio, Carrito, Perfil y Menú
 class Home extends StatelessWidget {
@@ -33,7 +34,7 @@ class Home extends StatelessWidget {
         case 'carrito':
           return Carrito();
         case 'perfil':
-          return Perfil();
+          return PerfilScreen();
         case 'menu': // Agregar este caso
           return Menu();
         default:
